Add "package.json" to exports map
Storybook telemetry uses the package.json to determine the installed version of packages. Because the Chromatic CLI now has an exports map, but it does not contain package.json as an entry point, Storybook fails to read it, and therefore the version in the metadata is `null`. This PR fixes that!