:sparkles: Add functionality to parse arrays and literal blocks for skip.
The action previously only allowed skip to be a boolean, string, or
empty. It now allows skip to contain boolean, string, an array of
strings, or a block literal. If an array or block literal is specified,
it will utilize these as though they are OR components of the glob
selector.

Fixes CAP-3601.